pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-repro-agent"
version = "0.1.0"
description = "Multi-subagent framework for automated clinical paper reproduction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "PyMuPDF>=1.24.0",
  "pdfplumber>=0.11.0",
]

[project.scripts]
paper-repro = "repro_agent.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
